:recycle: Refactor codes for shiny tray control.
* Make it works under `Electron 13.1.4`
