+++ src/assets/icons_for_shiny_tray.js
