+++ src/pages/shared/window-buttons/window-buttons.js
@@ -82,6 +82,7 @@ function checkUrlParams() {
     let title = params.get('title')
     let hide = params.get('hide')
     let trusted = params.get('trusted')
+    let script = params.get('script')
 
     if (trusted) {
         webview.classList.add('hide')
@@ -90,6 +91,16 @@ function checkUrlParams() {
     }
 
     if (page) webview.src = `../../${page}.html`
+
+    if (script) {
+        script = script.split(',')
+
+        script.forEach((src) => {
+            script_block = document.createElement('script')
+            script_block.src = `./${src}.js`
+            document.body.append(script_block)
+        })
+    }
 
     if (icon) {
         let elIcon = document.getElementById('icon')
